Create a basic step-by-step guide for setting up a video link for FPV
I've attempted to create a basic document to make things a little easier for new users.  The idea is to take them through the flashing of the camera and ground station, setting up the wifi adapters, and configuring the software.

This will get them to getting a basic video stream working.  There's more to come when I document the rest of the FPV setup